backend1/requirements.txt
# Workspace version: 1.42.49
databutton==0.38.34
fastapi==0.111.0
uvicorn[standard]==0.29.0
ipykernel==6.29.5
python-multipart==0.0.9
openai
beautifulsoup4
requests
google-generativeai